Streamline checks for minimum index compatibility
Today we have checks for the minimum version we support that are
not automatically updated. This change streamlines these checks
and makes them hard-coded by default to N-1 in a single place.
This also adds support to open SegmentInfos with a different
major version thatn N-1 in order to allow opening previous indices
given that codecs and postings lists are available via SPI.